# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TC API",
        "contact": {
            "name": "Trailercentral",
            "email": "alberto@trailercentral.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/website/parts/filters": {
            "get": {
                "tags": [
                    "Website Part Filters"
                ],
                "description": "Retrieve a list of filters",
                "operationId": "App\\Http\\Controllers\\v1\\Website\\Parts\\FilterController::index",
                "responses": {
                    "200": {
                        "description": "Returns a list of parts",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Error: Bad request."
                    }
                }
            }
        }
    }
}